[update] coderefair and AddCode
코드 리팩토링을 실시했습니다. 그래프 조회부분을 아에 싹다고쳤고 엔티티명도 바꾸었습니다.
기존에 tag에서 postSource라는 명으로 변경했습니다. 추후 hashtag계발을 위해 변경했습니다.
그래프 조회를 부분조회 가능하도록 바꾸었습니다.
매핑테이블을 키로 묶어서 조회하면 쿼리가 더 적게 발생할거 같습니다.
JWT가 추가됨에 따라 시큐리티에 토큰인증이 필요한곳과 안필요한곳 경로를 지정했습니다.
또한 컨트롤러에서 토큰의 정보가 필요한 부분에 토큰을 받도록 코드를 수정했습니다.

+++ capstone2023/src/main/java/com/capstone/domain/post/service/PostService.java
@@ -27,12 +27,13 @@ import com.capstone.domain.log.service.LogService;
 import com.capstone.domain.post.dto.PostRequest;
 import com.capstone.domain.post.dto.PostResponse;
 import com.capstone.domain.post.entity.Post;
+import com.capstone.domain.post.exception.PostForbiddenException;
 import com.capstone.domain.post.exception.PostNotFoundException;
 import com.capstone.domain.post.mapper.PostMapper;
 import com.capstone.domain.post.repository.PostRepository;
+import com.capstone.domain.postSource.service.PostSourceService;
 import com.capstone.domain.reply.dto.ReplyResponse;
 import com.capstone.domain.reply.mapper.ReplyMapper;
-import com.capstone.domain.tag.service.TagService;
 import com.capstone.domain.user.entity.User;
 import com.capstone.domain.user.exception.UserNotFoundException;
 import com.capstone.domain.user.repository.UserRepository;
@@ -48,7 +49,7 @@ public class PostService {
 
 	private final FileService fileService;
 	private final LogService logService;
-	private final TagService tagService;
+	private final PostSourceService tagService;
 	
 	private final PostMapper postMapper;
 	private final LogMapper logMapper;
@@ -79,7 +80,7 @@ public class PostService {
 		//재창작인 경우 
 		
 		if(board.getBno()==4&&postRequest.getTag() !=null) {
-			tagService.tagCreate(postRequest.getTag(),post);
+			tagService.postSourceCreate(postRequest.getTag(),post);
 		}	
 	}
 	
@@ -137,7 +138,11 @@ public class PostService {
 			Post post= this.postRepository.findByPno(pno).orElseThrow(() -> new PostNotFoundException()) ;
 			//pk값은 존재하고 나머지 값이 null일 경우에 nullpointException을 추가적으로 발행해줘야함.
 			User user = this.userRepository.findByUno(uno).orElseThrow(() -> new UserNotFoundException());
-			Board board = this.boardRepository.findByBno(postDTO.getBno()).orElseThrow( () -> new BoardNotFoundException());		
+			Board board = this.boardRepository.findByBno(postDTO.getBno()).orElseThrow( () -> new BoardNotFoundException());
+			if(!post.getUser().getUno().equals(user.getUno())) {
+				System.out.println("post : "+ post.getUser().getUno()+" user : " + user.getUno());
+				throw new PostForbiddenException();
+			}
 			post = postMapper.toEntity(postDTO ,board ,user);
 			return postMapper.toPostResponse(this.postRepository.save(post));
 	}
@@ -145,7 +150,11 @@ public class PostService {
 	
 	//게시판 삭제메소드
 	@Transactional
-	public void postDelete(Long pno) {
+	public void postDelete(Long pno, Long uno) {
+		Post post= this.postRepository.findByPno(pno).orElseThrow(() -> new PostNotFoundException()) ;
+		User user = this.userRepository.findByUno(uno).orElseThrow(() -> new UserNotFoundException());
+		if(!post.getUser().getUno().equals(user.getUno()))
+			throw new PostForbiddenException();
 		this.postRepository.deleteById(pno);
 	}
