+++ capstone2023/src/main/java/com/capstone/domain/reply/service/ReplyService.java
@@ -1,22 +1,23 @@
 package com.capstone.domain.reply.service;
 
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.capstone.domain.post.entity.Post;
+import com.capstone.domain.post.exception.PostNotFoundException;
+import com.capstone.domain.post.repository.PostRepository;
 import com.capstone.domain.reply.dto.ReplyRequest;
 import com.capstone.domain.reply.entity.Reply;
+import com.capstone.domain.reply.exception.ReplyForbiddenException;
 import com.capstone.domain.reply.exception.ReplyNotFoundException;
 import com.capstone.domain.reply.exception.ReplyNullPointException;
 import com.capstone.domain.reply.mapper.ReplyMapper;
 import com.capstone.domain.reply.repository.ReplyRepository;
-import com.capstone.domain.post.entity.Post;
-import com.capstone.domain.post.exception.PostNotFoundException;
-import com.capstone.domain.post.repository.PostRepository;
 import com.capstone.domain.user.entity.User;
 import com.capstone.domain.user.exception.UserNotFoundException;
 import com.capstone.domain.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
+import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Service
@@ -30,10 +31,10 @@ public class ReplyService {
 
 
     @Transactional
-    public void replyCreate(ReplyRequest replyDTO) {
+    public void replyCreate(ReplyRequest replyDTO,Long uno) {
 
         // 유저 참조 설정
-        User user = this.userRepository.findByUno(replyDTO.getUno())
+        User user = this.userRepository.findByUno(uno)
                 .orElseThrow(() -> new UserNotFoundException());
 
         // 게시글 참조 설정
@@ -53,25 +54,27 @@ public class ReplyService {
     }
 
     @Transactional
-    public void replyDelete(Long rno) {
+    public void replyDelete(Long rno,Long uno) {
         Reply reply = (Reply) replyRepository.findByRno(rno)
                 .orElseThrow(() -> new ReplyNullPointException());
-
+        if(!reply.getUser().getUno().equals(uno)) {
+        	throw new ReplyForbiddenException();
+        }
         // 댓글 삭제 로직
         replyRepository.delete(reply);
     }
 
     @Transactional
-    public void replyUpdate(ReplyRequest replyDTO) {
+    public void replyUpdate(ReplyRequest replyDTO,Long uno) {
 
         // 유저 참조 설정
-        User user = this.userRepository.findByUno(replyDTO.getUno())
+        User user = this.userRepository.findByUno(uno)
                 .orElseThrow(() -> new UserNotFoundException());
 
         // 게시글 참조 설정
         Post post = this.postRepository.findByPno(replyDTO.getPno())
                 .orElseThrow(() -> new PostNotFoundException());
-
+        
         // 부모 댓글 참조 설정(대댓글일 경우)
         Reply parentReply = null;
         if (replyDTO.getDeph() == 2) {
@@ -80,6 +83,9 @@ public class ReplyService {
         }
         // 댓글 수정 로직
         Reply reply = replyMapper.toEntity(replyDTO, user, post, parentReply);
+        if(!reply.getUser().getUno().equals(uno)) {
+        	throw new ReplyForbiddenException();
+        }
         replyRepository.save(reply);
     }
 }
